add: add Get-methods with stats

+++ statsApp/src/main/java/ru/practicum/stat/model/ViewStats.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class ViewStats {
     String app;
     String uri;
-    Long hits;
+    Integer hits;
 }
